+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/implementation/aad/configuration/RestTemplateProxyCustomizerTestConfiguration.java
@@ -13,7 +13,7 @@ import java.net.InetSocketAddress;
 import java.net.Proxy;
 
 @Configuration
-public class RestTemplateProxyCustomizerConfiguration {
+public class RestTemplateProxyCustomizerTestConfiguration {
 
     public static final SimpleClientHttpRequestFactory FACTORY = createProxyFactory();